more fix on define-struct
Corrected the macro definitions
so that in
```
(define-struct s (x))
(define-struct (t s) (w y))
```
only s-x and t-w, t-y are defined
and arity of `make-t` is corrected
as well.

ToDo: the parent word in 'struct
is always #f now, as can't find a way
to save the location struct is stored
in define-struct-info.