potential bug: args in mutator-func definition aren't roots
This is to meant to fix the bug that the args of self-defined functions within plai/gc2/mutator.rkt are not treated like
other stack/env roots. So in some rare edge case these args
would be lost by collection.

The best solution to this problem is to make every args of self-
defined functions root belongs to the active-roots list. Concern
is when shall the active-roots list be cleared. The current strategy
is to clear it after every root/loc retrieval, but not sure if that's
safe enough.