Assembled all collectors
Partially because they are valid implementations which
should not be deprecated or moved to trash. Partially
because collectors w/o profiling built-in are more suitable
for complicated mutator debugging.

Some not-ready updates on locality profiling was mistakenly
committed as well. Will correct/optimize it in the near future.

Not all collectors support 'vector, 'struct and 'struct-instance.
Need future update.